Add hashes to most static asset file names to invalidate cache
I noticed that my browser was using a stale cached version of `icons.svg`, resulting in some icons on the site not being displayed correctly. This was due to incorrect configuration of the static file server. While it was correctly including an ETag in the response, it wasn’t sending cache headers to instruct browsers to actually revalidate files using the ETag.

Instead of fixing this, I decided to change the set up so that we import static files directly in the components that use them. During the build, esbuild will automatically copy these files to the `dist` directory and append a hash based on the file contents. For example, `icons.svg` becomes `icons-JUG7VFSS.svg`. When the file changes, its URL changes, and browsers will never fetch and outdated cached version.

As an added benefit, this also ensures that the build fails when we reference a static file that doesn’t exist.

There are some files where this doesn’t work, for example the SVG logos of the political groups. We’d have to explicitly import all of these which would be kind of annoying. But they rarely change, so I think that’s fine for now.
