+++ frontend/src/css/fonts.css
@@ -1,8 +1,8 @@
 @font-face {
   font-family: "IBM Plex Sans Condensed";
   src:
-    url("/static/fonts/ibm-plex-sans-condensed-semibold.woff2") format("woff2"),
-    url("/static/fonts/ibm-plex-sans-condensed-semibold.woff") format("woff");
+    url("../fonts/ibm-plex-sans-condensed-semibold.woff2") format("woff2"),
+    url("../fonts/ibm-plex-sans-condensed-semibold.woff") format("woff");
   font-weight: 600;
   font-style: normal;
   font-display: swap;
@@ -11,8 +11,8 @@
 @font-face {
   font-family: "IBM Plex Sans";
   src:
-    url("/static/fonts/ibm-plex-sans-semibold.woff2") format("woff2"),
-    url("/static/fonts/ibm-plex-sans-semibold.woff") format("woff");
+    url("../fonts/ibm-plex-sans-semibold.woff2") format("woff2"),
+    url("../fonts/ibm-plex-sans-semibold.woff") format("woff");
   font-weight: 600;
   font-style: normal;
   font-display: swap;
@@ -21,8 +21,8 @@
 @font-face {
   font-family: "IBM Plex Sans";
   src:
-    url("/static/fonts/ibm-plex-sans-medium.woff2") format("woff2"),
-    url("/static/fonts/ibm-plex-sans-medium.woff") format("woff");
+    url("../fonts/ibm-plex-sans-medium.woff2") format("woff2"),
+    url("../fonts/ibm-plex-sans-medium.woff") format("woff");
   font-weight: 500;
   font-style: normal;
   font-display: swap;
@@ -31,8 +31,8 @@
 @font-face {
   font-family: "IBM Plex Sans";
   src:
-    url("/static/fonts/ibm-plex-sans-regular.woff2") format("woff2"),
-    url("/static/fonts/ibm-plex-sans-regular.woff") format("woff");
+    url("../fonts/ibm-plex-sans-regular.woff2") format("woff2"),
+    url("../fonts/ibm-plex-sans-regular.woff") format("woff");
   font-weight: 400;
   font-style: normal;
   font-display: swap;
