+++ frontend/src/custom.d.ts
@@ -1,0 +1,14 @@
+declare module "*.png" {
+  const url: string;
+  export default url;
+}
+
+declare module "*.jpeg" {
+  const url: string;
+  export default url;
+}
+
+declare module "*.svg" {
+  const url: string;
+  export default url;
+}
