+++ frontend/src/fonts/ibm-plex-sans-regular.woff2
